requirements.txt
streamlit
streamlit-pdf-viewer
pdfplumber
sentence-transformers
transformers
torch
numpy
tiktoken
langdetect
sentencepiece
accelerate
groq
PyPDF2